supprimer les fichiers uploads + ajouter package fs + ajouter un modal confirm avant suppression

+++ frontend/src/pages/admin/user/GestionUserAdmin.jsx
@@ -2,7 +2,7 @@ import React, { useEffect, useState } from "react";
 import MoreVertIcon from "@mui/icons-material/MoreVert";
 import AddIcon from "@mui/icons-material/Add";
 //import avatar from "@img/profile.png";
-import { NavLink, useNavigate } from "react-router-dom";
+import { NavLink } from "react-router-dom";
 import EditNoteIcon from "@mui/icons-material/EditNote";
 import HighlightOffIcon from "@mui/icons-material/HighlightOff";
 import VisibilityIcon from "@mui/icons-material/Visibility";
@@ -10,17 +10,18 @@ import userService from "@services/user.service";
 const GestionUserAdmin = () => {
   const [users, setUsers] = useState([]);
   const [loading, setLoading] = useState(false);
-  const navigate = useNavigate();
+
 
   //suppression d'un utilisateur
 
   const handleDeteleUserAdmin = (id) => {
-    userService.deleteUserAdmin(id).then((res) => {
-      console.log(res.data.message);
-      navigate("/admin/gestion/user");
-    });
-
-    console.log("delete");
+    if (window.confirm("Voulez vous supprimer cet utilisateur")) {
+      userService.deleteUserAdmin(id).then((res) => {
+        console.log(res.data.message);
+        setUsers(users.filter(user => user.id !== id))
+    
+      });
+    }
   };
 
   useEffect(() => {
@@ -55,20 +56,20 @@ const GestionUserAdmin = () => {
     );
 
   return (
-    <div class="relative overflow-x-auto shadow-md">
+    <div className="relative overflow-x-auto shadow-md">
       <div className="bg-gray-50 flex flex-row justify-between mb-3 p-5">
         <div>
-          <form class="max-w-md mx-auto">
+          <form className="max-w-md mx-auto">
             <label
-              for="default-search"
-              class="mb-2 text-sm font-medium text-gray-900 sr-only dark:text-white"
+              htmlFor="default-search"
+              className="mb-2 text-sm font-medium text-gray-900 sr-only dark:text-white"
             >
               Search
             </label>
-            <div class="relative">
-              <div class="absolute inset-y-0 start-0 flex items-center ps-3 pointer-events-none">
+            <div className="relative">
+              <div className="absolute inset-y-0 start-0 flex items-center ps-3 pointer-events-none">
                 <svg
-                  class="w-4 h-4 text-gray-500 dark:text-gray-400"
+                  className="w-4 h-4 text-gray-500 dark:text-gray-400"
                   aria-hidden="true"
                   xmlns="http://www.w3.org/2000/svg"
                   fill="none"
@@ -76,9 +77,9 @@ const GestionUserAdmin = () => {
                 >
                   <path
                     stroke="currentColor"
-                    stroke-linecap="round"
-                    stroke-linejoin="round"
-                    stroke-width="2"
+                    strokeLinecap="round"
+                    strokeLinejoin="round"
+                    strokeWidth="2"
                     d="m19 19-4-4m0-7A7 7 0 1 1 1 8a7 7 0 0 1 14 0Z"
                   />
                 </svg>
@@ -86,7 +87,7 @@ const GestionUserAdmin = () => {
               <input
                 type="search"
                 id="default-search"
-                class="block w-full p-2 ps-10 text-sm text-gray-900 border border-gray-300  bg-gray-50 focus:ring-blue-500 focus:border-blue-500 dark:bg-gray-700 dark:border-gray-600 dark:placeholder-gray-400 dark:text-white dark:focus:ring-blue-500 dark:focus:border-blue-500"
+                className="block w-full p-2 ps-10 text-sm text-gray-900 border border-gray-300  bg-gray-50 focus:ring-blue-500 focus:border-blue-500 dark:bg-gray-700 dark:border-gray-600 dark:placeholder-gray-400 dark:text-white dark:focus:ring-blue-500 dark:focus:border-blue-500"
                 placeholder="Rechercher..."
                 required
               />
@@ -106,34 +107,34 @@ const GestionUserAdmin = () => {
         </NavLink>
       </div>
 
-      <table class="w-full text-sm text-left rtl:text-right text-gray-500 dark:text-gray-400">
-        <thead class="text-xs text-[#00598a] uppercase bg-[#ecfeff] dark:bg-gray-700 dark:text-gray-400">
+      <table className="w-full text-sm text-left rtl:text-right text-gray-500 dark:text-gray-400">
+        <thead className="text-xs text-[#00598a] uppercase bg-[#ecfeff] dark:bg-gray-700 dark:text-gray-400">
           <tr>
-            <th scope="col" class="p-4">
-              <div class="flex items-center">Voir</div>
+            <th scope="col" className="p-4">
+              <div className="flex items-center">Voir</div>
             </th>
 
-            <th scope="col" class="px-6 py-3">
+            <th scope="col" className="px-6 py-3">
               Avatar
             </th>
 
-            <th scope="col" class="px-6 py-3">
+            <th scope="col" className="px-6 py-3">
               Nom d'utilisateur
             </th>
-            <th scope="col" class="px-6 py-3">
+            <th scope="col" className="px-6 py-3">
               Email
             </th>
-            <th scope="col" class="px-6 py-3">
+            <th scope="col" className="px-6 py-3">
               Prenom
             </th>
-            <th scope="col" class="px-6 py-3">
+            <th scope="col" className="px-6 py-3">
               Nom
             </th>
 
-            <th scope="col" class="px-6 py-3">
+            <th scope="col" className="px-6 py-3">
               Crée le
             </th>
-            <th scope="col" class="px-6 py-3">
+            <th scope="col" className="px-6 py-3">
               Actions
             </th>
           </tr>
@@ -144,10 +145,10 @@ const GestionUserAdmin = () => {
               return (
                 <tr
                   key={index}
-                  class="bg-white border-b dark:bg-gray-800 dark:border-gray-700 border-gray-200 hover:bg-gray-50 dark:hover:bg-gray-600"
+                  className="bg-white border-b dark:bg-gray-800 dark:border-gray-700 border-gray-200 hover:bg-gray-50 dark:hover:bg-gray-600"
                 >
-                  <td class="w-4 p-4">
-                    <div class="flex items-center">
+                  <td className="w-4 p-4">
+                    <div className="flex items-center">
                       <NavLink to={`/admin/gestion/user/show/${user.id}`}>
                         {" "}
                         <VisibilityIcon />{" "}
@@ -155,28 +156,28 @@ const GestionUserAdmin = () => {
                     </div>
                   </td>
 
-                  <td class="px-6 py-4">
+                  <td className="px-6 py-4">
                     {" "}
                     <img
-                      class=" w-8 h-8 rounded-full"
+                      className=" w-8 h-8 rounded-full"
                       src={`http://localhost:4000/uploads/avatars/${user.avatar}`}
                       alt="image description"
                     />
                   </td>
                   <th
                     scope="row"
-                    class="px-6 py-4 font-medium text-gray-900 whitespace-nowrap dark:text-white"
+                    className="px-6 py-4 font-medium text-gray-900 whitespace-nowrap dark:text-white"
                   >
                     {user.username || "N/A"}
                   </th>
 
-                  <td class="px-6 py-4"> {user.email || "N/A"}</td>
-                  <td class="px-6 py-4">{user.firstname || "N/A"}</td>
-                  <td class="px-6 py-4">{user.lastname || "N/A"}</td>
-                  <td class="px-6 py-4">
+                  <td className="px-6 py-4"> {user.email || "N/A"}</td>
+                  <td className="px-6 py-4">{user.firstname || "N/A"}</td>
+                  <td className="px-6 py-4">{user.lastname || "N/A"}</td>
+                  <td className="px-6 py-4">
                     {new Date(user.createdAt).toLocaleDateString() || "N/A"}
                   </td>
-                  <td class="px-6 py-4">
+                  <td className="px-6 py-4">
                     <NavLink onClick={() => handleDeteleUserAdmin(user.id)}>
                       <HighlightOffIcon className="cursor-pointer mr-3 " />
                     </NavLink>
@@ -192,20 +193,20 @@ const GestionUserAdmin = () => {
         </tbody>
       </table>
       <nav
-        class="flex items-center flex-column flex-wrap md:flex-row justify-between p-4"
+        className="flex items-center flex-column flex-wrap md:flex-row justify-between p-4"
         aria-label="Table navigation"
       >
-        <span class="text-sm font-normal text-gray-500 dark:text-gray-400 mb-4 md:mb-0 block w-full md:inline md:w-auto">
+        <span className="text-sm font-normal text-gray-500 dark:text-gray-400 mb-4 md:mb-0 block w-full md:inline md:w-auto">
           Affichage
-          <span class="font-semibold text-gray-900 dark:text-white">1-10</span>
+          <span className="font-semibold text-gray-900 dark:text-white">1-10</span>
           de
-          <span class="font-semibold text-gray-900 dark:text-white">1000</span>
+          <span className="font-semibold text-gray-900 dark:text-white">1000</span>
         </span>
-        <ul class="inline-flex -space-x-px rtl:space-x-reverse text-sm h-8">
+        <ul className="inline-flex -space-x-px rtl:space-x-reverse text-sm h-8">
           <li>
             <a
               href="#"
-              class="flex items-center justify-center px-3 h-8 ms-0 leading-tight text-gray-500 bg-white border border-gray-300 rounded-s-lg hover:bg-gray-100 hover:text-gray-700 dark:bg-gray-800 dark:border-gray-700 dark:text-gray-400 dark:hover:bg-gray-700 dark:hover:text-white"
+              className="flex items-center justify-center px-3 h-8 ms-0 leading-tight text-gray-500 bg-white border border-gray-300 rounded-s-lg hover:bg-gray-100 hover:text-gray-700 dark:bg-gray-800 dark:border-gray-700 dark:text-gray-400 dark:hover:bg-gray-700 dark:hover:text-white"
             >
               précédent
             </a>
@@ -213,7 +214,7 @@ const GestionUserAdmin = () => {
           <li>
             <a
               href="#"
-              class="flex items-center justify-center px-3 h-8 leading-tight text-gray-500 bg-white border border-gray-300 hover:bg-gray-100 hover:text-gray-700 dark:bg-gray-800 dark:border-gray-700 dark:text-gray-400 dark:hover:bg-gray-700 dark:hover:text-white"
+              className="flex items-center justify-center px-3 h-8 leading-tight text-gray-500 bg-white border border-gray-300 hover:bg-gray-100 hover:text-gray-700 dark:bg-gray-800 dark:border-gray-700 dark:text-gray-400 dark:hover:bg-gray-700 dark:hover:text-white"
             >
               1
             </a>
@@ -221,7 +222,7 @@ const GestionUserAdmin = () => {
           <li>
             <a
               href="#"
-              class="flex items-center justify-center px-3 h-8 leading-tight text-gray-500 bg-white border border-gray-300 hover:bg-gray-100 hover:text-gray-700 dark:bg-gray-800 dark:border-gray-700 dark:text-gray-400 dark:hover:bg-gray-700 dark:hover:text-white"
+              className="flex items-center justify-center px-3 h-8 leading-tight text-gray-500 bg-white border border-gray-300 hover:bg-gray-100 hover:text-gray-700 dark:bg-gray-800 dark:border-gray-700 dark:text-gray-400 dark:hover:bg-gray-700 dark:hover:text-white"
             >
               2
             </a>
@@ -230,7 +231,7 @@ const GestionUserAdmin = () => {
             <a
               href="#"
               aria-current="page"
-              class="flex items-center justify-center px-3 h-8 text-blue-600 border border-gray-300 bg-blue-50 hover:bg-blue-100 hover:text-blue-700 dark:border-gray-700 dark:bg-gray-700 dark:text-white"
+              className="flex items-center justify-center px-3 h-8 text-blue-600 border border-gray-300 bg-blue-50 hover:bg-blue-100 hover:text-blue-700 dark:border-gray-700 dark:bg-gray-700 dark:text-white"
             >
               3
             </a>
@@ -238,7 +239,7 @@ const GestionUserAdmin = () => {
           <li>
             <a
               href="#"
-              class="flex items-center justify-center px-3 h-8 leading-tight text-gray-500 bg-white border border-gray-300 hover:bg-gray-100 hover:text-gray-700 dark:bg-gray-800 dark:border-gray-700 dark:text-gray-400 dark:hover:bg-gray-700 dark:hover:text-white"
+              className="flex items-center justify-center px-3 h-8 leading-tight text-gray-500 bg-white border border-gray-300 hover:bg-gray-100 hover:text-gray-700 dark:bg-gray-800 dark:border-gray-700 dark:text-gray-400 dark:hover:bg-gray-700 dark:hover:text-white"
             >
               4
             </a>
@@ -246,7 +247,7 @@ const GestionUserAdmin = () => {
           <li>
             <a
               href="#"
-              class="flex items-center justify-center px-3 h-8 leading-tight text-gray-500 bg-white border border-gray-300 hover:bg-gray-100 hover:text-gray-700 dark:bg-gray-800 dark:border-gray-700 dark:text-gray-400 dark:hover:bg-gray-700 dark:hover:text-white"
+              className="flex items-center justify-center px-3 h-8 leading-tight text-gray-500 bg-white border border-gray-300 hover:bg-gray-100 hover:text-gray-700 dark:bg-gray-800 dark:border-gray-700 dark:text-gray-400 dark:hover:bg-gray-700 dark:hover:text-white"
             >
               5
             </a>
@@ -254,7 +255,7 @@ const GestionUserAdmin = () => {
           <li>
             <a
               href="#"
-              class="flex items-center justify-center px-3 h-8 leading-tight text-gray-500 bg-white border border-gray-300 rounded-e-lg hover:bg-gray-100 hover:text-gray-700 dark:bg-gray-800 dark:border-gray-700 dark:text-gray-400 dark:hover:bg-gray-700 dark:hover:text-white"
+              className="flex items-center justify-center px-3 h-8 leading-tight text-gray-500 bg-white border border-gray-300 rounded-e-lg hover:bg-gray-100 hover:text-gray-700 dark:bg-gray-800 dark:border-gray-700 dark:text-gray-400 dark:hover:bg-gray-700 dark:hover:text-white"
             >
               suivant
             </a>
